requirements.txt
boto3==1.36.26
botocore==1.36.26
jmespath==1.0.1
py4j==0.10.9.7
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
urllib3==1.26.20
